content/media/test/test_audio1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: Audio Constructor Test 1</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="use_large_cache.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var testsWaiting = 0;

for (var i = 0; i < gAudioTests.length; ++i) {
  var test = gAudioTests[i];
  var a1 = new Audio();
  if (a1.canPlayType(test.type) == "no")
    continue;

  a1.setAttribute("autobuffer", "");
  a1.addEventListener('load', function(event) {
    is(event.target.networkState, HTMLMediaElement.NETWORK_LOADED,
       "Audio " + event.target.currentSrc + " loaded");
    --testsWaiting;
    if (testsWaiting == 0) {
      SimpleTest.finish();
    }
  }, false);
  a1.src = test.name;
  a1.load();
  ++testsWaiting;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
